emacs: reformat and modularize code blocks
This will be an ongoing process. Currently, the code blocks have
multiple packages being called and are a monolith. Splitting up the code
blocks so that I can add more commentary and follow literate programming
more. Plus, it should make it easier for future me to understand
decisions made on code used.