emacs: add org-mode info and move clocking time
This commit adds some comments on why Org-mode is being used and how to
use it. Most importantly, this also updates how the clocked-in time
header is being rendered in the status bar. Previously, if the heading
is too long, it will be removed entirely. So this change simply
truncates the heading.