emacs: more templates, contact, and task efforts
Most of this commit is house keeping the configuration file. The biggest
and most productive change is the one to ask for task estimates when
clocking in. This has been a nice reminder to how much work a task
should take before going on, which can increase a sense of urgency.